requirements.txt
Embedly==0.5.0
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
arrow==0.4.4
contextlib2==0.4.0
elasticsearch==1.2.0
futures==2.2.0
httplib2==0.9
itsdangerous==0.24
lucene-querybuilder==0.2
mock==1.0.1
porc==0.3.1
python-dateutil==2.2
redis==2.10.3
requests==2.4.3
requests-futures==0.9.4
rq==0.4.6
rq-dashboard==0.3.4
simplejson==3.6.4
six==1.8.0
times==0.7
urllib3==1.9.1
vcrpy==1.1.0
wsgiref==0.1.2
